refactor(game): read [skill_check:] tags through one shared parser
The client parsed a GM check tag properly and the server did not. The
server read `skill` and `dc` with its own attribute grep and dropped
everything else, so an advantage mode or a d20 the player rolled
themselves was thrown away the moment the server resolved the check.
The client also owned the only audit of the GM's own arithmetic, which
left the audit unavailable to any server-side resolver.

Moves `parseSkillCheckTagBody` (and its rolls reader) verbatim into
`packages/shared/src/utils/skill-check-tag.ts`, along with the
`SkillCheckTag` type and the tag grammar itself, so both sides read a
check the same way and reach the same verdict about whether the GM's
numbers can be trusted. The client re-exports the type; behavior is
unchanged on this commit.

The grammar is exported as a source string plus a
`createSkillCheckTagRegex()` factory rather than a shared regex object,
because a global regex carries `lastIndex` and sharing one instance
across readers would make matching order-dependent.
