+++ packages/shared/src/index.ts
@@ -124,6 +124,7 @@ export * from "./utils/music-score.js";
 export * from "./utils/agent-cost.js";
 export * from "./utils/regex-replacement.js";
 export * from "./utils/skill-check-format.js";
+export * from "./utils/skill-check-tag.js";
 export * from "./utils/generation-guide.js";
 export * from "./utils/lorebook-keyword-matching.js";
 export * from "./utils/regex-safety.js";
